Normal always outer + small correction
If normal to parametrised curvature is calculated as N=T'/|T|, with T tangent, the direction changes with the curvature. New implementation of normal N = (T1,-T0) with T=(T0,T1).
Moreover, there was a small error in the evaluation of the para_curve in cut not agglo cell: fixed.